requirements.txt
PyYAML
matplotlib
pandas==0.23.4
pyprind
sklearn
scipy
bayesian-optimization==0.6
tables
numpy==1.16.0
torch
